refactor: refactor page
