+++ packages/client/src/clientPages/home/index.ts
@@ -1,0 +1,1 @@
+export { ClientHomePage } from './ui/ClientHomePage'
