+++ packages/client/src/design-system/icon/default/index.ts
@@ -1,6 +1,8 @@
 export { IconAddPhoto } from './IconAddPhoto'
+export { IconAlarm } from './IconAlarm'
 export { IconArrowDown } from './IconArrowDown'
 export { IconArrowLeft } from './IconArrowLeft'
+export { IconArrowRight } from './IconArrowRight'
 export { IconArrowUp } from './IconArrowUp'
 export { IconBookmark } from './IconBookmark'
 export { IconBookmarkFilled } from './IconBookmarkFilled'
